Set the correct DIJKSTRA_MAX value

+++ include/sine_nomine/dijkstra.h
@@ -27,7 +27,7 @@ struct dijkstra_map {
 	dijkstra **values;
 };
 
-enum { DIJKSTRA_MAX = USHRT_MAX,
+enum { DIJKSTRA_MAX = UINT_MAX,
 };
 
 struct dijkstra_map *dijkstra_create(struct level *_level);
